% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouduseraccounts_objects.R
\name{AuthorizedKeysView}
\alias{AuthorizedKeysView}
\title{Cloud User Accounts API Objects 
Creates and manages users and groups for accessing Google Compute Engine virtual machines.}
\usage{
AuthorizedKeysView(keys = NULL, sudoer = NULL)
}
\arguments{
\item{keys}{[Output Only] The list of authorized public keys in SSH format}

\item{sudoer}{[Output Only] Whether the user has the ability to elevate on the instance that requested the authorized keys}
}
\value{
AuthorizedKeysView object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:34:47
filename: /Users/mark/dev/R/autoGoogleAPI/googleclouduseraccountsbeta.auto/R/clouduseraccounts_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
AuthorizedKeysView Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of authorized public keys for a user account.
}
