% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouduseraccounts_functions.R
\docType{package}
\name{clouduseraccounts_googleAuthR}
\alias{clouduseraccounts_googleAuthR}
\alias{clouduseraccounts_googleAuthR-package}
\title{Cloud User Accounts API
Creates and manages users and groups for accessing Google Compute Engine virtual machines.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:00:20
filename: /Users/mark/dev/R/autoGoogleAPI/googleclouduseraccountsbeta.auto/R/clouduseraccounts_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
\item https://www.googleapis.com/auth/cloud.useraccounts
\item https://www.googleapis.com/auth/cloud.useraccounts.readonly
}
}

