% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouduseraccounts_functions.R
\name{groups.addMember}
\alias{groups.addMember}
\title{Adds users to the specified group.}
\usage{
groups.addMember(GroupsAddMemberRequest, project, groupName)
}
\arguments{
\item{GroupsAddMemberRequest}{The \link{GroupsAddMemberRequest} object to pass to this method}

\item{project}{Project ID for this request}

\item{groupName}{Name of the group for this request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud.useraccounts
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/cloud.useraccounts)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/compute/docs/access/user-accounts/api/latest/}{Google Documentation}

Other GroupsAddMemberRequest functions: \code{\link{GroupsAddMemberRequest}}
}
